requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
exceptiongroup==1.2.2
frozendict==2.4.6
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.52
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
logging-utils==1.0.2
multitasking==0.0.11
numpy==2.2.4
ollama==0.4.7
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
openpyxl==3.1.2
peewee==3.17.9
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
ccxt==4.2.15
zstandard==0.23.0
matplotlib 
seaborn
yfinance
